Pytest - changes request PR - Struturing App
