+++ pi_server/app/__init__.py
@@ -1,3 +1,10 @@
 from flask import Flask
+from flask import redirect
 
-app = Flask("cat_bot")
+app = Flask(__name__)
+
+from app import routes_app
+
+
+
+
