+++ pi_server/test/addition/test_addition.py
@@ -1,4 +1,4 @@
-from app import add
+from app.addition.addition import add
 
 
 class TestAddition:
