+++ requirements.txt
@@ -1,1 +1,4 @@
-streamlit==1.7.0
+streamlit~=1.6.0
+setuptools~=58.1.0
+Flask~=2.0.3
+yagmail~=0.15.277
